Flight1-main/django_project/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
decorator==5.0.6
defusedxml==0.7.1
Django==3.2.3
django-bootstrap-form==3.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
docopt==0.6.2
entrypoints==0.3
FeinCMS==1.20.0
fileupload==0.1.5
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyperclip==1.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
yarg==0.1.9
